+++ tools/iceworks/app/main/services/customBlocks/index.js
@@ -1,4 +1,0 @@
-const dsl2code = require('./dsl2code');
-const getEngine = require('./getEngine');
-
-module.exports = { dsl2code, getEngine };